// src/app/models/registerRequest.model.ts
export class registerRequest {
  firstName: any;
  middleName: any;
  lastName: any;
  fatherName: any;
  motherName: any;
  address1: any;
  address2: any;
  occupation: any;
  dob: any;
  email: any;
  mobileNumber: any;
  referredCode: any;
  consent_id: any;
  ip_address: any;
  gender: any;
  agentType: any;
}